backend/requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.4.0
boto3==1.35.5
botocore==1.35.5
cachetools==4.2.1
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
drain3==0.9.11
ecdsa==0.19.0
exceptiongroup==1.2.2
fastapi==0.112.0
h11==0.14.0
idna==3.7
jmespath==1.0.1
jsonpickle==1.5.1
pyasn1==0.6.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
python-jose==3.3.0
requests==2.32.3
rsa==4.9
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.12.2
urllib3==1.26.19
uvicorn==0.30.6
